+++ examples/Commands/src/App.ts
@@ -282,14 +282,6 @@ namespace LiteMolPluginInstance {
         const theme = CustomTheme.createTheme(model.props.model, coloring);
         // instead of "polymer-visual", "model" or any valid ref can be used: all "child" visuals will be colored.
         CustomTheme.applyTheme(plugin, 'polymer-visual', theme); 
-
-        // in "real" applications, it is a good idea to cache the theme object.
-
-        // this coloring will be reset when the camera position is reset and needs to be updated manually
-        // by subscribing to the appropriate command:
-        //
-        //   plugin.subscribe(Bootstrap.Command.Visual.ResetScene, () => setTimeout(() => CustomTheme.applyTheme(plugin, 'polymer-visual', theme), 25));
-        //
     });
     
     addButton('Focus Query',  () => {