+++ CHANGELOG.md
@@ -5,6 +5,11 @@ This a global changelog that outlines overal changes in LiteMol.
 Each component also has its separate changelog 
 ([Core](src/lib/Core/CHANGELOG.md), [Visualization](src/lib/Visualization/CHANGELOG.md), [Bootstrap](src/lib/Bootstrap/CHANGELOG.md), [Plugin](src/lib/Plugin/CHANGELOG.md), [Viewer](src/Viewer/CHANGELOG.md)).
 
+2.2.1-preview
+-----------
+
+* Added support for "sticky" custom themes. See Commands example's CustomTheme for example usage.
+
 2.2.0-preview
 -----------